udiskslinuxfilesystem: Add udisks_linux_filesystem_is_volume_based()
This should be a common function to check whether the filesystem
is a traditional one, device number based, or whether it is a filesystem
of a different concept, based e.g. on 'volumes' such as btrfs, and
to indicate that any filesystem matching should be done some different
way, e.g. by filesystem UUID.