udiskslinuxfilesystem: Match mountpoints by ID_FS_UUID for volume-based filesystems
For filesystems with concept of 'volumes' the device number of an active
mount often doesn't match the real backing block device. A different
approach must be taken to match the mount with a block device. The
generic ID_FS_UUID udev attribute looks like a good candidate and
it seems to be widely used nowadays. For the moment this approach is
only used for btrfs devices.

The downside of such approach is grouping of all devices carrying
the particular attribute value under one volume, no matter whether
they're actually attached to an active mount or if it's just a copy
(duplicate UUIDs caused by full byte copy of a device).