udisksstate: Filesystem UUID support for mountpoint cleanup
This extends the on-disk mount state file structure with filesystem UUID
if available or requested. This is an initial support that might be
further extended to consider other block devices in the group
(filesystem volume or an active mountpoint). For now the cleanup takes
a more conservative approach in such case.